Try to fix merge issues
